Evaluate validation functions targeting two fields
Validation functions now contain declarations of all fields
of target entity (not nested, i.e. does not count with relations
yet). Also added support for rules that should be bound to
multiple attributes. At this moment, multiple functions are
created and bound to each attribute, but in the future, only
one function should be created, probably updated on submit.

Refs #32,39